pyproject.toml
[project]
name = "hajri-ai"
version = "0.1.0"
description = "A smart attendance system using face recognition."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "albumentations>=2.0.8",
    "deepface>=0.0.95",
    "fpdf2>=2.8.4",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.1.1",
    "setuptools<81",
    "streamlit>=1.50.0",
    "streamlit-authenticator>=0.4.2",
    "tf-keras>=2.20.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
